refactor: autoload CLI namespace and instance-based Processor
- Replace require_relative with Ruby autoload throughout lib/suma/;
  define parent-namespace autoload files (lib/suma/eengine.rb,
  lib/suma/cli/core.rb) so children lazy-load
- Split lib/suma/cli.rb into namespace + Cli::Core dispatcher; delegate
  subcommands to Cli::Build, Cli::Export, Cli::Compare, Cli::Validate,
  Cli::Reformat, Cli::GenerateSchemas, Cli::ExtractTerms,
  Cli::GenerateRegister, Cli::ConvertJsdai
- Convert Processor from class methods to instance-based with dependency
  injection via constructor kwargs (compile, output_directory,
  schemas_all_path)
- Utils.log now writes to $stderr by default with SUMA_DEBUG gating
- Add svg_quality/formatters.rb parent namespace for terminal/json/yaml
  formatters
