+++ exe/suma
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require "suma/cli"
+require "suma"
 
 Suma::Cli::Core.start(ARGV)
