+++ lib/suma/cli/validate.rb
@@ -9,8 +9,6 @@ module Suma
       desc "links SCHEMAS_FILE DOCUMENTS_PATH [OUTPUT_FILE]",
            "Extract and validate express links without creating intermediate file"
       def links(*args)
-        require_relative "validate_links"
-
         # Forward the command to ValidateLinks
         links = Cli::ValidateLinks.new
         links.extract_and_validate(*args)
