+++ lib/suma/eengine.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Suma
+  module Eengine
+    autoload :Errors,  "suma/eengine/errors"
+    autoload :Wrapper, "suma/eengine/wrapper"
+  end
+end
