+++ lib/suma/eengine/wrapper.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require "open3"
-require_relative "errors"
 
 module Suma
   module Eengine
