+++ lib/suma/jsdai.rb
@@ -1,6 +1,9 @@
+# frozen_string_literal: true
+
 module Suma
   module Jsdai
+    autoload :Figure,      "suma/jsdai/figure"
+    autoload :FigureImage, "suma/jsdai/figure_image"
+    autoload :FigureXml,   "suma/jsdai/figure_xml"
   end
 end
-
-require_relative "jsdai/figure"
